requirements.txt
weaviate-client==4.4.2
lxml==5.1.0
python-dotenv==1.0.1
requests==2.31.0
